Created one select instead game-select and region-select
