--- client/src/components/form/selects/region-select/style.module.css
@@ -1,13 +1,0 @@
-.caption {
-  --height: 20px;
-    overflow: hidden;
-    display: block;
-    height: var(--height);
-    line-height: var(--height);
-    font-size: 15px;
-    color: #fff;
-    white-space: nowrap;
-    text-overflow: ellipsis;
-    margin-bottom: 4px;
-    font-weight: bold;
-}
